pyproject.toml
[project]
name = "task-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Hiroyuki Kaneko",email = "neconoko.develop@gmail.com"}
]
packages = [
    { include = "task_assistant" }   # ← add this line
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "slack-sdk (>=3.35.0,<4.0.0)",
    "pytest-env (>=1.1.5,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
pytest = "^8.3.5"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
